requirements.txt
azure-identity==1.15.0
msgraph-sdk==1.0.0
python-dotenv==1.0.0
